add ManyToMany bound

+++ src/main/java/hexlet/code/app/service/LabelService.java
@@ -3,6 +3,7 @@ package hexlet.code.app.service;
 import hexlet.code.app.dto.LabelCreateDTO;
 import hexlet.code.app.dto.LabelDTO;
 import hexlet.code.app.dto.LabelUpdateDTO;
+import hexlet.code.app.exception.LabelInUseException;
 import hexlet.code.app.exception.ResourceNotFoundException;
 import hexlet.code.app.mapper.LabelMapper;
 import hexlet.code.app.repository.LabelRepository;
@@ -48,8 +49,13 @@ public class LabelService {
     }
 
     public void deleteLabel(Long id) {
-        labelRepository.findById(id)
+        var label = labelRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Label not found!"));
+
+        if (!label.getTasks().isEmpty()) {
+            throw new LabelInUseException("Label is used by tasks");
+        }
+
         labelRepository.deleteById(id);
     }
 }
